Simplify nutchindexing, and support hadoop1 and cdh5-mr2 now

+++ nutchindexing/bin/run.sh
@@ -29,20 +29,17 @@ if [ -n "$1" ]; then
 fi
 
 NUTCH_BIN_DIR=$HIBENCH_HOME"/common/hibench/nutchindexing/target"
-export COMMON_DEPENDENCY_DIR=$HIBENCH_HOME"/common/hibench/common/target/dependency"
-export NUTCHINDEXING_DEPENDENCY_DIR=$NUTCH_BIN_DIR"/dependency"
+COMMON_DEPENDENCY_DIR=$HIBENCH_HOME"/common/hibench/common/target/dependency"
 
 if [ ! -e $NUTCH_BIN_DIR"/apache-nutch-1.2-bin.tar.gz" ]; then
   echo "Error: The nutch bin file hasn't be downloaded by maven, please check!"
   exit
 fi
 
-if [ $HADOOP_VERSION == "hadoop1" -a -e $DIR"/nutch/conf/nutch-site-mr1.xml" ]; then
-  mv $DIR/nutch/conf/nutch-site.xml $DIR/nutch/conf/nutch-site-mr2.xml
-  mv $DIR/nutch/conf/nutch-site-mr1.xml $DIR/nutch/conf/nutch-site.xml
-elif [ $HADOOP_VERSION == "hadoop2" -a -e $DIR"/nutch/conf/nutch-site-mr2.xml" ]; then
-  mv $DIR/nutch/conf/nutch-site.xml $DIR/nutch/conf/nutch-site-mr1.xml
-  mv $DIR/nutch/conf/nutch-site-mr2.xml $DIR/nutch/conf/nutch-site.xml
+if [ $HADOOP_VERSION == "hadoop1" ]; then
+  cp $DIR/nutch/conf/nutch-site-mr1.xml $DIR/nutch/conf/nutch-site.xml
+elif [ $HADOOP_VERSION == "hadoop2" ]; then
+  cp $DIR/nutch/conf/nutch-site-mr2.xml $DIR/nutch/conf/nutch-site.xml
 fi
 
 cd $NUTCH_BIN_DIR
@@ -51,23 +48,22 @@ if [ ! -d $NUTCH_BIN_DIR"/nutch-1.2" ]; then
 fi
 
 NUTCH_HOME=$NUTCH_BIN_DIR/nutch-1.2
-rm -rf $NUTCH_HOME/conf/*
-rm -rf $NUTCH_HOME/bin/*
+find $NUTCH_HOME/lib ! -name "lucene-*" -type f -exec rm -rf {} \;
 cp $DIR/nutch/conf/nutch-site.xml $NUTCH_HOME/conf
 cp $DIR/nutch/bin/nutch $NUTCH_HOME/bin
-mkdir $NUTCH_HOME/temp
-unzip -q $NUTCH_HOME/nutch-1.2.job -d $NUTCH_HOME/temp
-rm $NUTCH_HOME/temp/lib/jcl-over-slf4j-*.jar
-cp $COMMON_DEPENDENCY_DIR/jcl-over-slf4j-*.jar $NUTCH_HOME/temp/lib
-rm $NUTCH_HOME/nutch-1.2.job
-cd $NUTCH_HOME/temp
-zip -qr $NUTCH_HOME/nutch-1.2.job *
-cd $NUTCH_HOME
-rm -rf $NUTCH_HOME/temp
 
-if [ -d $NUTCH_BIN_DIR"/nutch-1.2/lib" ]; then
-  rm -rf $NUTCH_BIN_DIR"/nutch-1.2/lib"
+
+if [ $HADOOP_VERSION == "hadoop2" ]; then
+  mkdir $NUTCH_HOME/temp
+  unzip -q $NUTCH_HOME/nutch-1.2.job -d $NUTCH_HOME/temp
+  rm $NUTCH_HOME/temp/lib/jcl-over-slf4j-*.jar
+  cp $COMMON_DEPENDENCY_DIR/jcl-over-slf4j-*.jar $NUTCH_HOME/temp/lib
+  rm $NUTCH_HOME/nutch-1.2.job
+  cd $NUTCH_HOME/temp
+  zip -qr $NUTCH_HOME/nutch-1.2.job *
+  rm -rf $NUTCH_HOME/temp
 fi
+cd $NUTCH_HOME
 
 check_compress
